// src/types/data.types.ts
// Generated by https://quicktype.io

export interface Data {
  currentUser: User;
  productRequests: ProductRequest[];
}

export interface User {
  image: string;
  name: string;
  username: string;
}

export interface ProductRequest {
  id: number;
  title: string;
  category: Lowercase<Category>;
  upvotes: number;
  status: Status;
  description: string;
  comments?: Comment[];
}

export enum Category {
  ui = "UI",
  ux = "UX",
  enhancement = "Enhancement",
  bug = "Bug",
  feature = "Feature",
}

export enum Status {
  SUGGESTION = "suggestion",
  PLANNED = "planned",
  IN_PROGRESS = "in-progress",
  LIVE = "live",
}

export interface Comment {
  id: number;
  content: string;
  user: User;
  replies?: Reply[];
}

export interface Reply {
  content: string;
  replyingTo: string;
  user: User;
}